# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  resources.SubscribeUserResource:
    properties:
      community_id:
        example: 507f1f77bcf86cd799439012
        type: string
      role:
        enum:
        - member
        - admin
        - owner
        example: member
        type: string
      user_id:
        example: 507f1f77bcf86cd799439013
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - community_id
    - role
    type: object
  resources.SubscriptionCountResource:
    properties:
      community_id:
        example: 507f1f77bcf86cd799439012
        type: string
      count:
        example: 150
        type: integer
    type: object
  resources.SubscriptionListResource:
    properties:
      subscriptions:
        items:
          $ref: '#/definitions/resources.SubscriptionResource'
        type: array
      total:
        example: 150
        type: integer
    type: object
  resources.SubscriptionResource:
    properties:
      community_id:
        example: 507f1f77bcf86cd799439012
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      role:
        example: member
        type: string
      subscription_id:
        example: 507f1f77bcf86cd799439011
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user_id:
        example: 507f1f77bcf86cd799439013
        type: string
    type: object
  resources.UnsubscribeUserResource:
    properties:
      community_id:
        example: 507f1f77bcf86cd799439012
        type: string
      user_id:
        example: 507f1f77bcf86cd799439013
        type: string
    required:
    - community_id
    - user_id
    type: object
...
host: localhost
info:
  contact: {}
  description: Community management API with Kafka event processing
  title: Gommunity API
  version: "1.0"
paths:
...
  /api/v1/subscriptions:
    delete:
      consumes:
      - application/json
      description: Remove a user's subscription from a community. Users can unsubscribe
        themselves, or owner/admin can remove users.
      parameters:
      - description: Unsubscribe request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resources.UnsubscribeUserResource'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unsubscribe a user from a community
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: 'Subscribe a user to a community with a specific role. IMPORTANT:
        Self-subscriptions (following a community) always receive ''member'' role
        regardless of requested role. In public communities, users can only subscribe
        themselves. In private communities, owner/admin can add users by username
        and assign any role.'
      parameters:
      - description: Subscription request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resources.SubscribeUserResource'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.SubscriptionResource'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe a user to a community
      tags:
      - subscriptions
  /api/v1/subscriptions/communities/{community_id}:
    get:
      description: Get all subscriptions for a specific community with optional pagination
      parameters:
      - description: Community ID
        in: path
        name: community_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.SubscriptionListResource'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all subscriptions for a community
      tags:
      - subscriptions
  /api/v1/subscriptions/communities/{community_id}/count:
    get:
      description: Get the total number of subscriptions for a specific community
      parameters:
      - description: Community ID
        in: path
        name: community_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.SubscriptionCountResource'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get subscription count for a community
      tags:
      - subscriptions
  /api/v1/subscriptions/users/{user_id}/communities/{community_id}:
    get:
      description: Get a specific subscription for a user in a community
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Community ID
        in: path
        name: community_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.SubscriptionResource'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get subscription by user and community
      tags:
      - subscriptions
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token directly (Bearer prefix is optional).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"